+++ packages/vue-clamp/src/index.ts
@@ -1,12 +1,8 @@
-export { Clamp, LineClamp } from "./LineClamp.ts";
+export { LineClamp } from "./LineClamp.ts";
 export { InlineClamp } from "./InlineClamp.ts";
 export { WrapClamp } from "./WrapClamp.ts";
 
 export type {
-  ClampExposed,
-  ClampLocation,
-  ClampProps,
-  ClampSlotProps,
   InlineClampParts,
   InlineClampProps,
   InlineClampSplit,